\name{LOPART_interface}
\alias{LOPART_interface}
\title{LOPART interface}

\usage{LOPART_interface(input_data, 
    input_label_start, 
    input_label_end, 
    input_label_changes, 
    penalty)}
\arguments{
  \item{input_data}{
}
  \item{input_label_start}{
}
  \item{input_label_end}{
}
  \item{input_label_changes}{
}
  \item{penalty}{
}
}



\author{Toby Dylan Hocking}





